@@ -32,6 +32,7 @@ import { MemberEditResolver } from './_resolvers/member-edit.resolver';
 import { PreventUnsavedChangesGuard } from './_guards/prevent-unsaved-changes.guard';
 import { MemberListResolver } from '../app/_resolvers/member-list.resolver';
 import { ListResolver } from './_resolvers/lists.resolver';
+import { MessageListResolver } from './_resolvers/message-list.resolver';
 
 // export function tokenGetter(): string {
 //   return localStorage.getItem('token');
@@ -86,7 +87,8 @@ import { ListResolver } from './_resolvers/lists.resolver';
     MemberEditResolver,
     PreventUnsavedChangesGuard,
     MemberListResolver,
-    ListResolver
+    ListResolver,
+    MessageListResolver
   ],
   bootstrap: [AppComponent]
 })
